# server_go/docs/swagger.yaml
basePath: /
definitions:
  httpapi.AuthResponse:
    properties:
      message:
        type: string
      statusCode:
        type: integer
    type: object
  httpapi.HTTPValidationError:
    properties:
      detail:
        items:
          $ref: '#/definitions/httpapi.ValidationError'
        type: array
    type: object
  httpapi.RequestValidationError:
    properties:
      message:
        type: string
      statusCode:
        type: integer
    type: object
  httpapi.SearchResponse:
    properties:
      data:
        items:
          additionalProperties: {}
          type: object
        type: array
    type: object
  httpapi.ValidationError:
    properties:
      loc:
        items: {}
        type: array
      msg:
        type: string
      type:
        type: string
    type: object
host: huw.dk
info:
  contact: {}
  description: API for the WhoKnows search application
  title: WhoKnows API
  version: "1.0"
paths:
  /:
    get:
      description: Serves the main search page as HTML.
      produces:
      - text/html
      responses:
        "200":
          description: HTML page
          schema:
            type: string
      summary: Serve Root Page
      tags:
      - pages
  /api/login:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: Authenticate a user with username and password. Sets a session
        cookie on success.
      parameters:
      - description: Username
        in: formData
        name: username
        required: true
        type: string
      - description: Password
        in: formData
        name: password
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httpapi.AuthResponse'
        "422":
          description: Validation Error
          schema:
            $ref: '#/definitions/httpapi.HTTPValidationError'
      summary: Login
      tags:
      - auth
  /api/logout:
    get:
      description: Clear the session and log the user out.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httpapi.AuthResponse'
      summary: Logout
      tags:
      - auth
  /api/register:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: Create a new user account. Validates input and checks for duplicate
        usernames.
      parameters:
      - description: Username
        in: formData
        name: username
        required: true
        type: string
      - description: Email
        in: formData
        name: email
        required: true
        type: string
      - description: Password
        in: formData
        name: password
        required: true
        type: string
      - description: Password2
        in: formData
        name: password2
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httpapi.AuthResponse'
        "422":
          description: Validation Error
          schema:
            $ref: '#/definitions/httpapi.HTTPValidationError'
      summary: Register
      tags:
      - auth
  /api/search:
    get:
      description: Search wiki pages by title. Returns matching pages as JSON.
      parameters:
      - description: Search query
        in: query
        name: q
        required: true
        type: string
      - description: Language code (e.g., 'en')
        in: query
        name: language
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httpapi.SearchResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/httpapi.RequestValidationError'
      summary: Search
      tags:
      - search
  /login:
    get:
      description: Serves the login page as HTML.
      produces:
      - text/html
      responses:
        "200":
          description: HTML page
          schema:
            type: string
      summary: Serve Login Page
      tags:
      - pages
  /register:
    get:
      description: Serves the registration page as HTML.
      produces:
      - text/html
      responses:
        "200":
          description: HTML page
          schema:
            type: string
      summary: Serve Register Page
      tags:
      - pages
swagger: "2.0"